add project endpoints
