+++ .gitignore
@@ -1,4 +1,4 @@
-project
+project/
 smarty/templates_c
 .gitmodules
 project-*
